+++ hutool-json/src/test/java/cn/hutool/json/JSONUtilTest.java
@@ -112,4 +112,14 @@ public class JSONUtilTest {
 		JSONObject jsonObject = JSONUtil.parseObj(html);
 		Assert.assertEquals("Something\\u00a0must\\u00a0have\\u00a0been\\u00a0changed\\u00a0since\\u00a0you\\u00a0leave", jsonObject.getStr("name"));
 	}
+	
+	@Test
+	public void parseFromXmlTest() {
+		String s = "<sfzh>640102197312070614</sfzh><sfz>640102197312070614X</sfz><name>aa</name><gender>1</gender>";
+		JSONObject json = JSONUtil.parseFromXml(s);
+		Assert.assertEquals(640102197312070614L, json.get("sfzh"));
+		Assert.assertEquals("640102197312070614X", json.get("sfz"));
+		Assert.assertEquals("aa", json.get("name"));
+		Assert.assertEquals(1, json.get("gender"));
+	}
 }